split models.py up
