+++ decklist/models/__init__.py
@@ -1,0 +1,18 @@
+from django.contrib.auth.models import AbstractUser
+
+
+class User(AbstractUser):
+    pass
+
+
+from .datasource import DataSource
+from .partnertype import PartnerType
+from .deck import Deck
+from .card import Card
+from .printing import Printing
+from .cardindeck import CardInDeck
+from .sitestat import SiteStat
+from .commander import Commander
+from .theme import Theme
+from .themeresult import ThemeResult
+from .synergyscore import SynergyScore
